feat(client): enhance Roadmap component with task status and priority indicators
